+++ src/components/Tooltip/Tooltip.module.scss
@@ -1,4 +1,4 @@
-.rc-tooltip-inner {
+.rcTooltipInner {
   text-transform: initial;
   font-weight: 400 !important;
 }
