+++ src/pages/DashboardPage/components/GraceList/GraceList.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react';
 import { NavLink } from 'react-router-dom';
 
 import { GraceListItem } from '@frakt/state/liquidations/types';
+import { createTimerJSX } from '@frakt/components/Timer';
 import Button from '../../../../components/Button';
-import { createTimerJSX } from '@frakt/utils';
 import { PATHS } from '../../../../constants';
 import styles from './GraceList.module.scss';
 import { Timer } from '../../../../icons';
