+++ src/pages/LiquidationsPage/components/GraceCard/GraceCard.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 
 import { GeneralCardInfo, StatsRaffleValues } from '../StatsRaffleValues';
 import { GraceListItem } from '@frakt/state/liquidations/types';
-import { createTimerJSX } from '@frakt/utils';
+import { createTimerJSX } from '@frakt/components/Timer';
 import styles from './GraceCard.module.scss';
 import { Timer } from '@frakt/icons';
