+++ src/pages/LiquidationsPage/components/RafflesList/RafflesList.module.scss
@@ -27,7 +27,7 @@
     left: 0;
     right: 0;
     top: unset;
-    width: 100% !important;
+    width: 100%;
   }
 }
